Step pattern editor (#164)
* Add StepPatternEditor shared VSTGUI control for TranceGate pattern editing

Implement a full-featured visual step pattern editor as a shared VSTGUI
CControl for editing TranceGate gate patterns with 2-32 steps. The
component is plugin-agnostic and communicates via ParameterCallback.

Key features:
- Click-and-drag vertical bar editing with paint mode across steps
- Fine mode (Shift+drag, 0.1x sensitivity) for precise ghost notes
- Euclidean rhythm generation with Bjorklund algorithm and dot indicators
- Real-time playback position indicator via atomic pointer polling
- Phase offset visualization, quick action presets, and transforms
- Scroll/zoom for high step counts (24+)
- 8 configurable color attributes via VSTGUI ViewCreator system

New files:
- plugins/shared/src/ui/step_pattern_editor.h (shared control)
- plugins/shared/src/ui/color_utils.h (shared color utilities)
- plugins/shared/tests/ (44 tests, 336 assertions)

Modified for integration:
- trance_gate_params.h: v2 state format with 32 step levels, Euclidean
  params, phase offset; v1 migration support
- controller.cpp: verifyView wiring, setParamNormalized sync, action
  button routing, playback poll timer
- processor.cpp: step level application, playback position IMessage
- editor.uidesc: StepPatternEditor view, action buttons, Euclidean toolbar
- plugin_ids.h: action button tags (10000-10009)

All 37 FR and 8 SC requirements verified. Pluginval passes at level 5.


* Add StepPatternEditor to control testbench and fix warnings

Add the StepPatternEditor control to the standalone testbench app with
parameter logging, 16-step default, and interactive instructions panel.
Fix CCoord-to-float narrowing warnings in TapPatternEditor and remove
unused srcPos variable in EuclideanPattern.


* Add plugins/shared change detection to CI workflow

Changes to shared UI components (ArcKnob, StepPatternEditor,
FieldsetContainer) now trigger rebuilds for all consuming plugins
(Iterum, Disrumpo, Ruinae) across all three platforms.


---------

+++ plugins/iterum/src/ui/tap_pattern_editor.cpp
@@ -60,14 +60,14 @@ void TapPatternEditor::drawGridLines(VSTGUI::CDrawContext* context) {
 
     // Vertical grid lines at 1/4 intervals (stop at ruler)
     for (int i = 1; i < 4; ++i) {
-        float x = viewRect.left + width * (static_cast<float>(i) / 4.0f);
+        float x = static_cast<float>(viewRect.left) + width * (static_cast<float>(i) / 4.0f);
         context->drawLine(
             VSTGUI::CPoint(x, viewRect.top),
             VSTGUI::CPoint(x, tapAreaBottom));
     }
 
     // Horizontal grid line at 50% level (within tap area)
-    float y = viewRect.top + tapAreaHeight * 0.5f;
+    float y = static_cast<float>(viewRect.top) + tapAreaHeight * 0.5f;
     context->drawLine(
         VSTGUI::CPoint(viewRect.left, y),
         VSTGUI::CPoint(viewRect.right, y));
@@ -88,8 +88,8 @@ void TapPatternEditor::drawTaps(VSTGUI::CDrawContext* context) {
         float level = tapLevels_[i];
 
         // Calculate tap bar position (within tap area, above ruler)
-        float centerX = viewRect.left + timeRatio * width;
-        float barTop = viewRect.top + (1.0f - level) * tapAreaHeight;
+        float centerX = static_cast<float>(viewRect.left) + timeRatio * width;
+        float barTop = static_cast<float>(viewRect.top) + (1.0f - level) * tapAreaHeight;
         float barBottom = static_cast<float>(viewRect.bottom) - kRulerHeight;
 
         // Tap bar rectangle
